astinterp: Implement ClassDef.
This in particular required tighten up how functions are represented.
Previously, it was just an object with __call__, but such an object
is not treated by Python as a "fully native function" is some contexts,
e.g. instance method binding. Solve that by wrapping the object
describe with a real function, dispatching to the object's __call__.

(A possible alternative was to producing bound methods on the AST
interpreter side, checking arguments in visit_Attribute, and transforming
them as needed. But Pycopy current doesn't export bound method constructor,
and overall, due to astinterp's simplicity goals, better to rely on the
underlying runtime, as in other places.)