astinterp: Fully switch to scope representation as linked list of namespaces.
It doesn't make sense to have 2 parallel representations: as explicit array
stack and as a linked list, this would only lead to confusion and mistakes.
Linked list representation is more flexible, because to store a scope stack,
we need to just store reference to the lowest namespace, whereas with explicit
stack, we would need to copy it.